Updated persistent login feature
