+++ src/App.js
@@ -1,4 +1,8 @@
+import { useEffect } from 'react';
+import Axios from 'axios';
+import { API_URL } from './assets/constants';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { useDispatch } from 'react-redux';
 
 import Home from './pages/Home';
 import Login from './pages/Login';
@@ -11,6 +15,38 @@ import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
 function App() {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    const persistentLogin = async () => {
+      const userToken = localStorage.getItem('userToken');
+      let response;
+
+      if (userToken) {
+        response = await Axios.get(`${API_URL}/auth/persistent`, {
+          headers: {
+            Authorization: `Bearer ${userToken}`,
+          },
+        });
+      } else {
+        response = await Axios.get(`${API_URL}/auth/login/success`, { withCredentials: true });
+      }
+
+      if (response.data.ignore) {
+        return;
+      } else {
+        localStorage.setItem('userToken', response.data.token);
+
+        dispatch({
+          type: 'USER_LOGIN',
+          payload: response.data.user,
+        });
+      }
+    };
+
+    persistentLogin();
+  }, []);
+
   return (
     <BrowserRouter>
       <Routes>
